+++ config/config.exs
@@ -124,7 +124,8 @@ config :td_perms,
     :view_rejected_ingests,
     :view_deprecated_ingests,
     :manage_confidential_structures,
-    :manage_ingest_relations
+    :manage_ingest_relations,
+    :view_data_structures_profile
   ]
 
 # Import environment specific config. This must remain at the bottom